Add example target for visionOS.
PiperOrigin-RevId: 566636331
(cherry picked from commit 0891f9a28234446d45d013eefec8a358f22bf47c)

# Conflicts:
#	examples/visionos/HelloWorld/BUILD
#	examples/visionos/HelloWorld/Info.plist
