+++ examples/resources/BUILD
@@ -24,6 +24,11 @@ filegroup(
 )
 
 filegroup(
+    name = "VisionAppIcon.xcassets",
+    srcs = glob(["VisionAppIcon.xcassets/**"]),
+)
+
+filegroup(
     name = "WatchAppIcon.xcassets",
     srcs = glob(["WatchAppIcon.xcassets/**"]),
 )
