+++ examples/resources/VisionAppIcon.xcassets/AppIcon.solidimagestack/Back.solidimagestacklayer/Content.imageset/Contents.json
@@ -1,0 +1,13 @@
+{
+  "images" : [
+    {
+      "filename" : "Bazel_logo.png",
+      "idiom" : "reality",
+      "scale" : "2x"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
