--- examples/visionos/HelloWorld/Sources/App.swift
@@ -1,16 +1,0 @@
-import SwiftUI
-
-@main
-struct VisionOSApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
-
-struct ContentView: View {
-    var body: some View {
-        Text("Hello World")
-    }
-}
